+++ apps/web/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @locusai/web
 
+## 0.9.16
+
+### Patch Changes
+
+- Sprints commands on telegram
+- Updated dependencies
+  - @locusai/sdk@0.9.16
+  - @locusai/shared@0.9.16
+
+## 0.9.15
+
+### Patch Changes
+
+- Default pull rebase logic
+- Updated dependencies
+  - @locusai/sdk@0.9.15
+  - @locusai/shared@0.9.15
+
+## 0.9.14
+
+### Patch Changes
+
+- Git configuration
+- Updated dependencies
+  - @locusai/sdk@0.9.14
+  - @locusai/shared@0.9.14
+
 ## 0.9.13
 
 ### Patch Changes